relocate shared infrastructure
