+++ src/shared/infrastructure/virtualization/VirtualGridCalculator.ts
