+++ src/shared/infrastructure/virtualization/__tests__/VirtualGridCalculator.test.ts
